+++ extension/naming-cdi/src/main/java/module-info.java
@@ -36,8 +36,8 @@
  */
 module cloud.piranha.extension.naming.cdi {
     
-    exports cloud.piranha.extension.naming;
-    opens cloud.piranha.extension.naming;
+    exports cloud.piranha.extension.naming.cdi;
+    opens cloud.piranha.extension.naming.cdi;
     requires cloud.piranha.core.api;
     requires transitive cloud.piranha.naming.impl;
     requires transitive cloud.piranha.naming.thread;
